requirements.txt
langchain==0.1.20
langgraph==0.0.65
streamlit==1.32.0
sentence-transformers==2.2.2
chromadb==0.4.24
networkx==3.2.1
python-dotenv==1.0.1
groq==0.8.0
